refactor(platform): extract session routing middleware
